change version type to long since swagger api type is uint64, ref: https://github.com/moby/moby/blob/master/api/swagger.yaml#L1969

+++ src/main/java/com/github/dockerjava/api/command/UpdateSwarmNodeCmd.java
@@ -23,10 +23,10 @@ public interface UpdateSwarmNodeCmd extends SyncDockerCmd<Void> {
 
     UpdateSwarmNodeCmd withSwarmNodeSpec(SwarmNodeSpec swarmNodeSpec);
 
-    UpdateSwarmNodeCmd withVersion(@Nonnull Integer versionId);
+    UpdateSwarmNodeCmd withVersion(@Nonnull Long versionId);
 
     @CheckForNull
-    Integer getVersion();
+    Long getVersion();
 
     interface Exec extends DockerCmdSyncExec<UpdateSwarmNodeCmd, Void> {
     }
